Extract SavedSearchList component from SearchTypeMenu
