+++ src/pages/Search/SearchTypeMenu.tsx
@@ -1,21 +1,14 @@
-import {useIsFocused, useRoute} from '@react-navigation/native';
-import {accountIDSelector} from '@selectors/Session';
+import {useRoute} from '@react-navigation/native';
 import React, {useCallback, useContext, useLayoutEffect, useMemo, useRef} from 'react';
 import {View} from 'react-native';
 // eslint-disable-next-line no-restricted-imports
 import type {ScrollView as RNScrollView, ScrollViewProps} from 'react-native';
 import MenuItem from '@components/MenuItem';
-import type {MenuItemWithLink} from '@components/MenuItemList';
-import MenuItemList from '@components/MenuItemList';
-import {usePersonalDetails} from '@components/OnyxListItemProvider';
-import {useProductTrainingContext} from '@components/ProductTrainingContext';
 import {ScrollOffsetContext} from '@components/ScrollOffsetContextProvider';
 import ScrollView from '@components/ScrollView';
 import {useSearchActionsContext} from '@components/Search/SearchContext';
 import type {SearchQueryJSON} from '@components/Search/types';
 import Text from '@components/Text';
-import useDeleteSavedSearch from '@hooks/useDeleteSavedSearch';
-import useFeedKeysWithAssignedCards from '@hooks/useFeedKeysWithAssignedCards';
 import {useMemoizedLazyExpensifyIcons} from '@hooks/useLazyAsset';
 import useLocalize from '@hooks/useLocalize';
 import useOnyx from '@hooks/useOnyx';
@@ -24,19 +17,15 @@ import useSingleExecution from '@hooks/useSingleExecution';
 import useSuggestedSearchDefaultNavigation from '@hooks/useSuggestedSearchDefaultNavigation';
 import useThemeStyles from '@hooks/useThemeStyles';
 import {setSearchContext} from '@libs/actions/Search';
-import {mergeCardListWithWorkspaceFeeds} from '@libs/CardUtils';
 import Navigation from '@libs/Navigation/Navigation';
-import {getAllTaxRates} from '@libs/PolicyUtils';
-import {buildSearchQueryJSON, buildUserReadableQueryString, shouldSkipSuggestedSearchNavigation as shouldSkipSuggestedSearchNavigationForQuery} from '@libs/SearchQueryUtils';
-import type {SavedSearchMenuItem} from '@libs/SearchUIUtils';
-import {createBaseSavedSearchMenuItem, getItemBadgeText, getOverflowMenu as getOverflowMenuUtil} from '@libs/SearchUIUtils';
+import {shouldSkipSuggestedSearchNavigation as shouldSkipSuggestedSearchNavigationForQuery} from '@libs/SearchQueryUtils';
+import {getItemBadgeText} from '@libs/SearchUIUtils';
 import variables from '@styles/variables';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import ROUTES from '@src/ROUTES';
 import todosReportCountsSelector from '@src/selectors/Todos';
-import type {SaveSearchItem} from '@src/types/onyx/SaveSearch';
-import SavedSearchItemThreeDotMenu from './SavedSearchItemThreeDotMenu';
+import SavedSearchList from './SavedSearchList';
 import SuggestedSearchSkeleton from './SuggestedSearchSkeleton';
 
 type SearchTypeMenuProps = {
@@ -52,20 +41,9 @@ function SearchTypeMenu({queryJSON}: SearchTypeMenuProps) {
     const {translate} = useLocalize();
     const [savedSearches] = useOnyx(ONYXKEYS.SAVED_SEARCHES);
     const {typeMenuSections, CreateReportConfirmationModal, shouldShowSuggestedSearchSkeleton} = useSearchTypeMenuSections();
-    const isFocused = useIsFocused();
-    const {
-        shouldShowProductTrainingTooltip: shouldShowSavedSearchTooltip,
-        renderProductTrainingTooltip: renderSavedSearchTooltip,
-        hideProductTrainingTooltip: hideSavedSearchTooltip,
-    } = useProductTrainingContext(
-        CONST.PRODUCT_TRAINING_TOOLTIP_NAMES.RENAME_SAVED_SEARCH,
-        !!typeMenuSections.find((section) => section.translationPath === 'search.savedSearchesMenuItemTitle') && isFocused,
-    );
     const expensifyIcons = useMemoizedLazyExpensifyIcons([
         'Basket',
-        'Bookmark',
         'CalendarSolid',
-        'Pencil',
         'Receipt',
         'ChatBubbles',
         'MoneyBag',
@@ -76,18 +54,7 @@ function SearchTypeMenu({queryJSON}: SearchTypeMenuProps) {
         'User',
         'Folder',
     ] as const);
-    const {showDeleteModal} = useDeleteSavedSearch();
-    const [allPolicies] = useOnyx(ONYXKEYS.COLLECTION.POLICY);
-    const personalDetails = usePersonalDetails();
-    const [cardList] = useOnyx(ONYXKEYS.CARD_LIST);
-    const [workspaceCardList] = useOnyx(ONYXKEYS.COLLECTION.WORKSPACE_CARDS_LIST);
-    const [reports] = useOnyx(ONYXKEYS.COLLECTION.REPORT);
-    const [allFeeds] = useOnyx(ONYXKEYS.COLLECTION.SHARED_NVP_PRIVATE_DOMAIN_MEMBER);
-    const feedKeysWithCards = useFeedKeysWithAssignedCards();
-    const taxRates = getAllTaxRates(allPolicies);
-    const [currentUserAccountID = -1] = useOnyx(ONYXKEYS.SESSION, {selector: accountIDSelector});
     const {clearSelectedTransactions} = useSearchActionsContext();
-    const cardsForSavedSearchDisplay = useMemo(() => mergeCardListWithWorkspaceFeeds(workspaceCardList ?? CONST.EMPTY_OBJECT, cardList), [workspaceCardList, cardList]);
     const [reportCounts = CONST.EMPTY_TODOS_REPORT_COUNTS] = useOnyx(ONYXKEYS.DERIVED.TODOS, {selector: todosReportCountsSelector});
 
     const flattenedMenuItems = useMemo(() => typeMenuSections.flatMap((section) => section.menuItems), [typeMenuSections]);
@@ -100,81 +67,7 @@ function SearchTypeMenu({queryJSON}: SearchTypeMenuProps) {
         shouldSkipNavigation: shouldSkipSuggestedSearchNavigation,
     });
 
-    const getOverflowMenu = useCallback(
-        (itemName: string, itemHash: number, itemQuery: string) => getOverflowMenuUtil(expensifyIcons, itemName, itemHash, itemQuery, translate, showDeleteModal),
-        [translate, showDeleteModal, expensifyIcons],
-    );
-    const createSavedSearchMenuItem = useCallback(
-        (item: SaveSearchItem, key: string, index: number) => {
-            let title = item.name;
-            if (title === item.query) {
-                const jsonQuery = buildSearchQueryJSON(item.query) ?? ({} as SearchQueryJSON);
-                title = buildUserReadableQueryString({
-                    queryJSON: jsonQuery,
-                    PersonalDetails: personalDetails,
-                    reports,
-                    taxRates,
-                    cardList: cardsForSavedSearchDisplay,
-                    cardFeeds: allFeeds,
-                    policies: allPolicies,
-                    currentUserAccountID,
-                    autoCompleteWithSpace: false,
-                    translate,
-                    feedKeysWithCards,
-                });
-            }
-
-            const isItemFocused = Number(key) === hash;
-            const baseMenuItem: SavedSearchMenuItem = createBaseSavedSearchMenuItem(item, key, index, title, isItemFocused);
-
-            return {
-                ...baseMenuItem,
-                sentryLabel: CONST.SENTRY_LABEL.SEARCH.SAVED_SEARCH_MENU_ITEM,
-                onPress: () => {
-                    setSearchContext(false);
-                    Navigation.navigate(ROUTES.SEARCH_ROOT.getRoute({query: item?.query ?? '', name: item?.name}));
-                },
-                rightComponent: (
-                    <SavedSearchItemThreeDotMenu
-                        menuItems={getOverflowMenu(title, Number(key), item.query)}
-                        isDisabledItem={item.pendingAction === CONST.RED_BRICK_ROAD_PENDING_ACTION.DELETE}
-                        hideProductTrainingTooltip={index === 0 && shouldShowSavedSearchTooltip ? hideSavedSearchTooltip : undefined}
-                        shouldRenderTooltip={index === 0 && shouldShowSavedSearchTooltip}
-                        renderTooltipContent={renderSavedSearchTooltip}
-                    />
-                ),
-                style: [styles.alignItemsCenter],
-                tooltipAnchorAlignment: {
-                    horizontal: CONST.MODAL.ANCHOR_ORIGIN_HORIZONTAL.RIGHT,
-                    vertical: CONST.MODAL.ANCHOR_ORIGIN_VERTICAL.BOTTOM,
-                },
-                tooltipShiftHorizontal: variables.savedSearchShiftHorizontal,
-                tooltipShiftVertical: variables.savedSearchShiftVertical,
-                tooltipWrapperStyle: [styles.mh4, styles.pv2, styles.productTrainingTooltipWrapper],
-                renderTooltipContent: renderSavedSearchTooltip,
-            };
-        },
-        [
-            hash,
-            getOverflowMenu,
-            shouldShowSavedSearchTooltip,
-            hideSavedSearchTooltip,
-            renderSavedSearchTooltip,
-            styles.alignItemsCenter,
-            styles.mh4,
-            styles.pv2,
-            styles.productTrainingTooltipWrapper,
-            personalDetails,
-            reports,
-            taxRates,
-            cardsForSavedSearchDisplay,
-            allFeeds,
-            feedKeysWithCards,
-            currentUserAccountID,
-            allPolicies,
-            translate,
-        ],
-    );
+    const isSavedSearchActive = !!savedSearches && Object.keys(savedSearches).some((key) => Number(key) === hash);
 
     const route = useRoute();
     const scrollViewRef = useRef<RNScrollView>(null);
@@ -198,42 +91,6 @@ function SearchTypeMenu({queryJSON}: SearchTypeMenuProps) {
         }
         scrollViewRef.current.scrollTo({y: scrollOffset, animated: false});
     }, [getScrollOffset, route]);
-
-    const {savedSearchesMenuItems, isSavedSearchActive} = useMemo(() => {
-        let savedSearchFocused = false;
-
-        if (!savedSearches) {
-            return {
-                isSavedSearchActive: false,
-                savedSearchesMenuItems: [],
-            };
-        }
-
-        const menuItems = Object.entries(savedSearches).map(([key, item], index) => {
-            const baseMenuItem = createSavedSearchMenuItem(item, key, index);
-            savedSearchFocused ||= !!baseMenuItem.focused;
-            return baseMenuItem;
-        });
-
-        return {
-            savedSearchesMenuItems: menuItems,
-            isSavedSearchActive: savedSearchFocused,
-        };
-    }, [createSavedSearchMenuItem, savedSearches]);
-
-    const renderSavedSearchesSection = useCallback(
-        (menuItems: MenuItemWithLink[]) => (
-            <MenuItemList
-                menuItems={menuItems}
-                wrapperStyle={styles.sectionMenuItem}
-                icon={expensifyIcons.Bookmark}
-                iconWidth={variables.iconSizeNormal}
-                iconHeight={variables.iconSizeNormal}
-                shouldUseSingleExecution
-            />
-        ),
-        [expensifyIcons.Bookmark, styles.sectionMenuItem],
-    );
 
     const activeItemIndex = useMemo(() => {
         // If we have a suggested search, then none of the menu items are active
@@ -268,7 +125,7 @@ function SearchTypeMenu({queryJSON}: SearchTypeMenuProps) {
                                 </Text>
 
                                 {section.translationPath === 'search.savedSearchesMenuItemTitle' ? (
-                                    renderSavedSearchesSection(savedSearchesMenuItems)
+                                    <SavedSearchList hash={hash} />
                                 ) : (
                                     <>
                                         {section.menuItems.map((item, itemIndex) => {
